requirements.txt
gradio==5.49.1
mineru_vl_utils==0.1.15
pdf2image==1.17.0
Pillow==11.1.0
transformers==4.57.1
accelerate==1.11.0